chore: Move unit tests and remove imports
- Move unit tests to proper directory away from integration tests
- Remove redundant import of `Setup` and `db` from tests
- Fix structure and documentation of running tests
