+++ app/api/helpers/scheduled_jobs.py
@@ -157,7 +157,7 @@ def expire_pending_tickets():
 
 
 def delete_ticket_holders_no_order_id():
-    from app import current_app as app
+    from app.instance import current_app as app
     with app.app_context():
         order_expiry_time = get_settings()['order_expiry_time']
         TicketHolder.query.filter(TicketHolder.order_id == None, TicketHolder.deleted_at.is_(None),
