+++ tests/all/integration/api/helpers/test_events.py
@@ -4,7 +4,6 @@ from app.api.helpers.db import save_to_db, get_count
 from app.api.helpers.events import create_custom_forms_for_attendees
 from app.factories.event import EventFactoryBasic
 from app.models.custom_form import CustomForms
-from tests.all.integration.setup_database import Setup
 from tests.all.integration.utils import OpenEventTestCase
