+++ tests/all/integration/api/helpers/test_export_helpers.py
@@ -2,7 +2,6 @@ import unittest
 
 from tests.all.integration.utils import OpenEventTestCase
 from tests.all.integration.auth_helper import create_user
-from tests.all.integration.setup_database import Setup
 from app.api.helpers.export_helpers import create_export_job
 from app.factories.export_job import ExportJobFactory
 from app.factories.event import EventFactoryBasic
