+++ tests/all/integration/api/helpers/test_jwt.py
@@ -5,9 +5,7 @@ from flask_jwt_extended import create_access_token
 from app.api.helpers.jwt import jwt_authenticate, get_identity
 from app.factories.event import EventFactoryBasic
 from app.factories.user import UserFactory
-from app.models import db
 from app.api.helpers.db import save_to_db
-from tests.all.integration.setup_database import Setup
 from tests.all.integration.utils import OpenEventTestCase
