+++ tests/all/integration/api/helpers/test_order.py
@@ -12,7 +12,6 @@ from app.factories.order import OrderFactory
 from app.models.order import Order
 from app.api.helpers.db import save_to_db
 from app.api.attendees import get_sold_and_reserved_tickets_count
-from tests.all.integration.setup_database import Setup
 from tests.all.integration.utils import OpenEventTestCase
