+++ tests/all/integration/api/helpers/test_pentabarfxml.py
@@ -1,7 +1,6 @@
 import unittest
 from datetime import datetime
 
-from tests.all.integration.setup_database import Setup
 from tests.all.integration.utils import OpenEventTestCase
 from app.api.helpers.pentabarfxml import PentabarfExporter
 from xml.etree.ElementTree import fromstring, tostring
