+++ tests/all/integration/api/validation/test_sessions.py
@@ -6,43 +6,9 @@ from tests.all.integration.utils import OpenEventTestCase
 from app.api.helpers.exceptions import UnprocessableEntity
 from app.api.schema.sessions import SessionSchema
 from app.factories.session import SessionFactory
-from app.models import db
-from app.api.helpers.db import save_to_db
-from tests.all.integration.setup_database import Setup
 
 
 class TestSessionValidation(OpenEventTestCase):
-
-    def test_date_pass(self):
-        """
-        Sessions Validate Date - Tests if the function runs without an exception
-        :return:
-        """
-        schema = SessionSchema()
-        original_data = {
-            'data': {}
-        }
-        data = {
-            'starts_at': datetime(2099, 8, 4, 12, 30, 45).replace(tzinfo=timezone('UTC')),
-            'ends_at': datetime(2099, 9, 4, 12, 30, 45).replace(tzinfo=timezone('UTC'))
-        }
-        SessionSchema.validate_fields(schema, data, original_data)
-
-    def test_date_start_gt_end(self):
-        """
-        Sessions Validate Date - Tests if exception is raised when ends_at is before starts_at
-        :return:
-        """
-        schema = SessionSchema()
-        original_data = {
-            'data': {}
-        }
-        data = {
-            'starts_at': datetime(2099, 9, 4, 12, 30, 45).replace(tzinfo=timezone('UTC')),
-            'ends_at': datetime(2099, 8, 4, 12, 30, 45).replace(tzinfo=timezone('UTC'))
-        }
-        with self.assertRaises(UnprocessableEntity):
-            SessionSchema.validate_fields(schema, data, original_data)
 
     def test_date_db_populate(self):
         """
@@ -51,8 +17,7 @@ class TestSessionValidation(OpenEventTestCase):
         """
         with self.app.test_request_context():
             schema = SessionSchema()
-            obj = SessionFactory()
-            save_to_db(obj)
+            SessionFactory()
 
             original_data = {
                 'data': {
