+++ tests/all/integration/setup_database.py
@@ -4,7 +4,6 @@ import sys
 import logging
 
 from app.instance import current_app as app
-from app.api.helpers.tasks import celery
 from app.models import db
 from app.models.setting import Environment
 from app.settings import set_settings
