+++ tests/all/unit/api/validation/test_discount_codes.py
@@ -1,11 +1,61 @@
 import unittest
+from unittest import TestCase
 
-from tests.all.integration.utils import OpenEventTestCase
 from app.api.helpers.exceptions import UnprocessableEntity
 from app.api.schema.discount_codes import DiscountCodeSchemaTicket
 
 
-class TestDiscountCodeValidation(OpenEventTestCase):
+class TestDiscountCodeValidation(TestCase):
+
+    def test_quantity_pass(self):
+        """
+        Discount Code Validate Quantity - Tests if the function runs without an exception
+        :return:
+        """
+        schema = DiscountCodeSchemaTicket()
+        original_data = {
+            'data': {}
+        }
+        data = {
+            'min_quantity': 10,
+            'max_quantity': 20,
+            'tickets_number': 30
+        }
+        DiscountCodeSchemaTicket.validate_quantity(schema, data, original_data)
+
+    def test_quantity_min_gt_max(self):
+        """
+        Discount Code Validate Quantity - Tests if exception is raised when min_quantity greater than max
+        :return:
+        """
+        schema = DiscountCodeSchemaTicket()
+        original_data = {
+            'data': {}
+        }
+        data = {
+            'min_quantity': 20,
+            'max_quantity': 10,
+            'tickets_number': 30
+        }
+        with self.assertRaises(UnprocessableEntity):
+            DiscountCodeSchemaTicket.validate_quantity(schema, data, original_data)
+
+    def test_quantity_max_gt_tickets_number(self):
+        """
+        Discount Code Validate Quantity - Tests if exception is raised when max_quantity greater than ticket_number
+        :return:
+        """
+        schema = DiscountCodeSchemaTicket()
+        original_data = {
+            'data': {}
+        }
+        data = {
+            'min_quantity': 10,
+            'max_quantity': 30,
+            'tickets_number': 20
+        }
+        with self.assertRaises(UnprocessableEntity):
+            DiscountCodeSchemaTicket.validate_quantity(schema, data, original_data)
 
     def test_percent_value_lte_hundred(self):
         """
