+++ tests/all/unit/utils/test_schema_defaulting.py
@@ -1,4 +1,5 @@
 import unittest
+from unittest import TestCase
 
 from marshmallow_jsonapi import fields
 from marshmallow_jsonapi.flask import Schema
@@ -24,25 +25,23 @@ class TestSchema(Schema):
     field_with_default = fields.Boolean(required=True, default=False)
 
 
-class TestUtils:
+class TestUtils(TestCase):
 
     def test_patch_defaults_adds_defaults(self):
-        with self.app.test_request_context():
-            schema = TestSchema()
-            data = {
-                'field_without_default': 'value_field_without_default'
-            }
-            patched_data = patch_defaults(schema, data)
-            self.assertEqual(patched_data.get('field_with_default'), False)
+        schema = TestSchema()
+        data = {
+            'field_without_default': 'value_field_without_default'
+        }
+        patched_data = patch_defaults(schema, data)
+        self.assertEqual(patched_data.get('field_with_default'), False)
 
     def test_patch_defaults_leaves_other_fields_untouched(self):
-        with self.app.test_request_context():
-            schema = TestSchema()
-            data = {
-                'field_without_default': 'value_field_without_default'
-            }
-            patched_data = patch_defaults(schema, data)
-            self.assertEqual(patched_data.get('field_without_default'), 'value_field_without_default')
+        schema = TestSchema()
+        data = {
+            'field_without_default': 'value_field_without_default'
+        }
+        patched_data = patch_defaults(schema, data)
+        self.assertEqual(patched_data.get('field_without_default'), 'value_field_without_default')
 
 
 if __name__ == '__main__':
